+++ buildSrc/src/main/kotlin/Dependencies.kt
@@ -26,7 +26,7 @@ object Dependencies {
     }
 
     object Android {
-        const val gradlePlugin = "com.android.tools.build:gradle:7.4.2"
+        const val gradlePlugin = "com.android.tools.build:gradle:8.0.2"
         const val compileSdkVersion = 34
         const val minSdkVersion = 21